First rough implementation of physics stuff
EMFoo is an 'environment' of sorts for ChargedFoos to interact. EMFoo
acts as the mediator and over-arching foo. ChargedFoos either repel or
attract each other, depending on their charge.

+++ EMFoo.h
@@ -17,7 +17,7 @@ class EMFoo : public Foo
 public:
 	EMFoo();
 	
-	int k;
+	double k;
 	
 	Force calculateForceOnFoo(Foo* aFoo);
